consumer/requirements.txt
 boto3
 pydantic-settings
 python-dotenv